% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSWPlib.R
\docType{package}
\name{LSWPlib}
\alias{LSWPlib}
\title{LSWPlib Package}
\description{
Library of functions for the statistical analysis and simulation of LSWP processes.
}
\author{
Alessandro Cardinali \email{alessandro.cardinali@plymouth.ac.uk}
}
