% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Daily log-returns for the S&P 500 stock index.}
\format{
A vector containing 1024 daily log-returns.
\describe{
  \item{Period Covered}{The data cover the period April 1996 - December 1999.}
}
}
\usage{
sp500
}
\description{
Log-returns from daily quotes of the Standard & Poors 500
Index.
}
\keyword{datasets}
